backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
python-multipart
email-validator
httpx
redis
